Changing all js to ts and setting strict mode to false
