+++ eslint.config.mjs
@@ -18,6 +18,7 @@ export default [
       "react/prop-types": "off",
       "no-unused-vars": "warn",
       "@typescript-eslint/no-unused-vars": "warn",
+      "@typescript-eslint/no-explicit-any": "warn",
     },
     settings: {
       react: {
