+++ package.json
@@ -11,7 +11,7 @@
     "lint": "eslint",
     "format": "prettier --write .",
     "check-format": "prettier --check .",
-    "lint-fix": "eslint . --fix --ext .js,.jsx,.ts,.tsx",
+    "lint-fix": "eslint . --fix",
     "dev": "next dev",
     "build": "next build",
     "start": "next start"
@@ -26,7 +26,6 @@
     "@radix-ui/react-select": "^2.1.2",
     "@radix-ui/react-slider": "^1.2.1",
     "@radix-ui/react-slot": "^1.1.0",
-    "@supabase/supabase-js": "^2.46.2",
     "@types/node": "^22.7.5",
     "@types/react": "^18.3.11",
     "aos": "^3.0.0-beta.6",
@@ -49,9 +48,9 @@
     "@babel/preset-env": "^7.26.0",
     "@babel/preset-react": "^7.25.9",
     "@babel/preset-typescript": "^7.26.0",
-    "@supabase/supabase-js": "^2.46.2",
     "@eslint/js": "^9.15.0",
     "@jest/globals": "^29.7.0",
+    "@supabase/supabase-js": "^2.46.2",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
