+++ tsconfig.json
@@ -4,7 +4,7 @@
     "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
-    "strict": true,
+    "strict": false,
     "forceConsistentCasingInFileNames": true,
     "noEmit": true,
     "esModuleInterop": true,
@@ -22,14 +22,16 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./*"]
-    }
+    },
+    "strictNullChecks": true
   },
   "include": [
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    ".next/types/**/*.ts"
-    // ".next/types/**/*.ts",
+    ".next/types/**/*.ts",
+    "utils/backendApi.js",
+    "utils/supabaseClient.js" // ".next/types/**/*.ts",
     // ".next/types/app/page.tsx",
     // ".next/types/app/layout.tsx",
     // ".next/types/app/(main)/layout.tsx",
